Split application to several files
